bochs: Send SIGSEGV instead of signal 0
Bochs always tries to find out the reason of an exception, so that
it can generate the right signal for gdb.

If it fails to find a reason, bochs assigns a value
GDBSTUB_STOP_NO_REASON (see bochs.h), which causes debug_loop()
(see gdbstub.cc) to generate a signal of number 0.

Signal 0 is problematic to gdb, as gdb doesn't allow us to ignore
it. Somehow when we simulate Linux, we get tons of signal 0's that
seem to be caused by page faults.

This chage makes bochs send SIGSEGV instead of signal 0, so that
we can ignore it in gdb.